Widen symbol lookup offset to full word width.
When symbol lookup fails, 64-bit systems may produce full 64-bit
deltas.  Use 'long' as the offset type to accommodate maximum possible
delta: vsyscall page at 0xffffffffff600000.  When printed out there
can be inappropriate negative numbers but this will get fixed in the
new dump format shortly.

+++ src/IgHookTrace.cc
@@ -25,8 +25,8 @@ bool
 IgHookTrace::symbol (void *address,
 		     const char *&sym,
 		     const char *&lib,
-		     int &offset,
-		     int &liboffset)
+		     long &offset,
+		     long &liboffset)
 {
     sym = lib = 0;
     offset = 0;
