+++ src/IgHookTrace.h
@@ -7,8 +7,8 @@ public:
     static int		stacktrace (void **addresses, int nmax);
     static void *	tosymbol (void *address);
     static bool		symbol (void *address, const char *&sym,
-		    		const char *&lib, int &offset,
-				int &liboffset);
+		    		const char *&lib, long &offset,
+				long &liboffset);
 };
 
 #endif // IG_HOOK_IG_HOOK_TRACE_H
